Aanpassingen gemaakt op het volgende: envirement omgezet naar Vite - Tevens de react-roter-dom aangepast naar nieuwe versie. Methodes useHistory en Switch omgezet naar useNavigate en Routes.
