--- src/App.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import { Switch, Route } from 'react-router-dom';
-import NavBar from './components/NavBar';
-import Profile from './pages/Profile';
-import Home from './pages/Home';
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
-import './App.css';
-
-function App() {
-  return (
-    <>
-      <NavBar />
-      <div className="content">
-        <Switch>
-          <Route exact path="/">
-            <Home />
-          </Route>
-          <Route path="/profile">
-            <Profile />
-          </Route>
-          <Route exact path="/signin">
-            <SignIn />
-          </Route>
-          <Route exact path="/signup">
-            <SignUp />
-          </Route>
-        </Switch>
-      </div>
-    </>
-  );
-}
-
-export default App;
