+++ src/App.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+import NavBar from './components/NavBar';
+import Profile from './pages/Profile';
+import Home from './pages/Home';
+import SignIn from './pages/SignIn';
+import SignUp from './pages/SignUp';
+import './App.css';
+
+function App() {
+  return (
+    <>
+      <NavBar />
+      <div className="content">
+        <Routes>
+
+          <Route 
+            exact path="/"
+            element={ <Home /> }
+          />
+
+          <Route 
+            path="/profile"
+            element={ <Profile /> }
+          />
+
+          <Route 
+            exact path="/signin"
+            element={ <SignIn /> }
+          />
+
+          <Route 
+            exact path="/signup"
+            element={ <SignUp /> }
+          />  
+        </Routes>
+      </div>
+    </>
+  );
+}
+
+export default App;
